requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
opencv-python
numpy
matplotlib
scikit-learn
Pillow
streamlit
streamlit-webrtc
